# public/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/games": {
            "get": {
                "tags": [
                    "Game"
                ],
                "summary": "Get list of all games",
                "description": "Get all games.",
                "operationId": "4778c61bb962d289c6f2a3fa7808e6fe",
                "responses": {
                    "200": {
                        "description": "Array of games",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Game"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Create a new game",
                "description": "Create a new game.",
                "operationId": "0c07edcb33a80a9c1160b2d534933e6c",
                "requestBody": {
                    "description": "Provide game name",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Chess"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Game created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/games/{id}": {
            "get": {
                "tags": [
                    "Game"
                ],
                "summary": "Get a single game",
                "description": "Get a single game by ID.",
                "operationId": "c16d89c6db36f6a71322534fe924e1e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Game"
                ],
                "summary": "Update an existing game",
                "description": "Update an existing game.",
                "operationId": "84b32a1acec4478423708cb806220cd3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Game data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Game"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Game"
                ],
                "summary": "Delete a game",
                "description": "Delete a game by ID.",
                "operationId": "9da75a65595bc905154e4685c7ec0a88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Game deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Game": {
                "title": "Game",
                "description": "Game model",
                "properties": {
                    "id": {
                        "description": "ID of the game",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nom": {
                        "description": "Name of the game",
                        "type": "string",
                        "example": "Chess"
                    }
                },
                "type": "object"
            },
...
        }
    }
}